+++ commands/__init__.py
@@ -1,4 +1,7 @@
 from .base import Base
 from .status import Status
+from .help import Help
+from .start import Start
+from .schedule import Schedule
 
-__all__ = ('Status')
+__all__ = ('Status', 'Start', 'Help', 'Schedule')