+++ main.py
@@ -7,26 +7,13 @@ from telegram import Bot
 from telegram.ext import Updater, CommandHandler
 from telegram.ext.filters import Filters
 
-from commands import Status
+from commands import Status, Help, Start, Schedule
 
 # Enable logging
 logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
                     level=logging.INFO)
 
 logger = logging.getLogger(__name__)
-
-
-# Define a few command handlers. These usually take the two arguments update and
-# context. Error handlers also receive the raised TelegramError object in error.
-def start(update, context):
-    """Send a message when the command /start is issued."""
-    logger.info(update.message['chat'])
-    update.message.reply_text(' 🍻')
-
-
-def help(update, context):
-    """Send a message when the command /help is issued."""
-    update.message.reply_text('Si no sabes como usarlo, no deberias estar jugando conmigo!!!')
 
 def error(update, context):
     """Log Errors caused by Updates."""
@@ -45,19 +32,16 @@ def main():
     # Get the dispatcher to register handlers
     dp = updater.dispatcher
 
-    # on different commands - answer in Telegram
-    dp.add_handler(CommandHandler("start", start, Filters.chat(-225411390) | Filters.chat(21667736)))
-    dp.add_handler(CommandHandler("help", help, Filters.chat(-225411390) | Filters.chat(21667736)))
-    dp.add_handler(Status.handler(updater.bot))
+    Start.setup(updater)
+    Help.setup(updater)
+    Status.setup(updater)
+    Schedule.setup(updater)
 
     # log all errors
     dp.add_error_handler(error)
 
     # Start the Bot
     updater.start_polling()
-
-    #
-    # bot.send_message(21667736, "Hola")
 
     logger.info("RepublicaBeerBot up and running")
     # Block until the user presses Ctrl-C or the process receives SIGINT,