+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## RepublicaBeerBot
 
+## 0.0.2
+
+* Refactor commands classes
+* Add `/help` command
+* Add `/schedule` command
+* Add `/start` command
+* Support to `add/show/remove` schedules on memory
+
 ## 0.0.1
 
 * Init telegram bot